Support setting `accessibilityElements`
Sometimes we want to override `accessibilityElements` so that we can
skip over elements unnecessary for interacting with the app using VoiceOver
or limit the interaction to child viewcontrollers presented, etc.

For this purpose, this commit adds support for setting custom
`accessibilityElements`. If not set, `accessibilityElements` should still
return the elements found from the view/layer hierarchy.