Support modifying/overriding value returned by `accessibilityElements`
Sometimes we want to override `accessibilityElements` so that we can
skip over elements unnecessary for interacting with the app using VoiceOver
or limit the interaction to child viewcontrollers presented, etc.

For this purpose, this commit introduces `accessibilityElementsBlock`
that receives the default `accessibilityElements` as input and modifies the
elements returned. If not set, `accessibilityElements` should still return
the elements found from the node hierarchy.